requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dash==3.2.0
dash-bootstrap-components==2.0.3
Flask==3.1.1
future==1.0.0
grpcio==1.59.0
grpcio-tools==1.59.0
idna==3.10
importlib_metadata==8.7.0
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
narwhals==2.1.1
nest-asyncio==1.6.0
packaging==25.0
plotly==6.3.0
protobuf==4.25.4
pyserial==3.5
PyYAML==6.0.2
requests==2.32.4
retrying==1.4.2
serial==0.0.97
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.23.0
